optimization: 优化 pipeline_post_save_handler 的实现, 改为异步方式进行相关数据统计
