+++ Procfile
@@ -1,5 +1,5 @@
 web: gunicorn wsgi -w 8 -b :$PORT --access-logfile - --error-logfile - --access-logformat '[%(h)s] %({request_id}i)s %(u)s %(t)s "%(r)s" %(s)s %(D)s %(b)s "%(f)s" "%(a)s"' --max-requests=200 --log-level debug --timeout 120
-dworker: python manage.py celery worker -Q default -n default@%h -c 6 -l info --maxtasksperchild=50
+dworker: celery worker -A blueapps.core.celery -P gevent -Q default -n default@%h -c 100 -l info --maxtasksperchild=50
 pworker: python manage.py celery worker -Q pipeline,pipeline_priority -n pipeline_worker@%h -c 10 -l info --maxtasksperchild=50
 sworker: celery worker -A blueapps.core.celery -P gevent -Q service_schedule,service_schedule_priority -c 100 -l info -n schedule_worker@%h --maxtasksperchild=50
 cworker: python manage.py celery worker -Q pipeline_additional_task,pipeline_additional_task_priority -n common_worker@%h -c 6 -l info --maxtasksperchild=50
